refactor left over files to directvolumemigrationprogress
