+++ pkg/controller/controller.go
@@ -19,8 +19,8 @@ package controller
 import (
 	"github.com/konveyor/mig-controller/pkg/controller/directimagemigration"
 	"github.com/konveyor/mig-controller/pkg/controller/directimagestreammigration"
-	"github.com/konveyor/mig-controller/pkg/controller/directpvmigrationprogress"
 	"github.com/konveyor/mig-controller/pkg/controller/directvolumemigration"
+	"github.com/konveyor/mig-controller/pkg/controller/directvolumemigrationprogress"
 	"github.com/konveyor/mig-controller/pkg/controller/discovery"
 	"github.com/konveyor/mig-controller/pkg/controller/miganalytic"
 	"github.com/konveyor/mig-controller/pkg/controller/migcluster"
@@ -49,7 +49,7 @@ var CamControllers = []AddFunction{
 	migplan.Add,
 	miganalytic.Add,
 	directvolumemigration.Add,
-	directpvmigrationprogress.Add,
+	directvolumemigrationprogress.Add,
 }
 
 //
