+++ pkg/controller/directvolumemigrationprogress/directpvmigrationprogress_controller_suite_test.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package directpvmigrationprogress
+package directvolumemigrationprogress
 
 import (
 	stdlog "log"
