+++ pkg/controller/directvolumemigrationprogress/directvolumemigrationprogress_controller.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package directpvmigrationprogress
+package directvolumemigrationprogress
 
 import (
 	"context"
@@ -72,7 +72,7 @@ func newReconciler(mgr manager.Manager) reconcile.Reconciler {
 // add adds a new Controller to mgr with r as the reconcile.Reconciler
 func add(mgr manager.Manager, r reconcile.Reconciler) error {
 	// Create a new controller
-	c, err := controller.New("directpvmigrationprogress-controller", mgr, controller.Options{Reconciler: r})
+	c, err := controller.New("directvolumemigrationprogress-controller", mgr, controller.Options{Reconciler: r})
 	if err != nil {
 		return err
 	}
